Remove generics from access control package (#35)
* Revert "chore(version): bump to 2.1.0"

This reverts commit 9c593f659a7a7de696d620e21b5dbf91f9508665.

* Revert "refactor(access): use generics for roles"

This reverts commit e25e91e9021410c5b729e5150ed7b05f0481561f.

* Revert "fix(ac): state CreateTypeSpec macro"

This reverts commit 3a10ad162af145b15ad79025163034b00e7059bb.

* chore(version): bump ac to 2.1.1

+++ contract/tests/cucumber.rs
@@ -348,7 +348,7 @@ fn mint_counts(world: &mut ContractWorld, user: String, count: u32) {
 #[then(regex = r"(\w+) user (has|has not) the (\w+) role")]
 fn user_is_admin(world: &mut ContractWorld, user: String, has: String, role: String) {
     let has_role = world.state.access_control.has_role(
-        get_user_role(role),
+        get_user_role(role) as u8,
         &mock_address(get_address_for_user(user)),
     );
